Escape spaces in paths.
If a path is passed in with a space in it from the bin_dir or any arguments, these spaces are not escaped. This causes the path no be cut off at that point.
This probably requires a bit of testing and checking it doesn't need to be added elsewhere.