Merge nested type extensions into original types
Replace a nested extension declaration (only valid at the top level)
with a skinny tree of extension declarations.  Matching the USR of
any existing type lets deduplication merge them all nicely.

Have to leave the declaration in the same position relative to the
others, rather than reparent it somewhere, to make sure the source-
code order is respected in the web pages.