Align 'jazzy' terminology with Apple usage
When comparing documentation generated by Jazzy with those on Apple's developer site, _The Swift Programming Language_, and documentation in the Swift Project GitHub repository, some of the terminology is written out differently in public-facing contexts. Since a consistent usage can be helpful for users, this proposed change aligns Jazzy with those canonical sources.

Hopefully, this is considered a helpful modification and not duplicative of any previous proposed changes :)