Fix Dash docset index creation
Prior to this Overview items were being represented in the Dash docset
index with a NULL type. Dash doesn't support indexes with NULL types in
them. To correct this I added an 'Overview' kind to the kind-type map
with a dash type of 'Section'. I talked about using this type with
Kapeli (the Dash developer) here,
https://github.com/Kapeli/Dash-User-Contributions/pull/1815 to see
if 'Section' was the appropriate type and it was.