Special-case deinit access level calculation
In Swift 5 SourceKit reports the accessibility of `deinit` with
no explicit access to be that of the enclosing type.  This commit
makes jazzy continue to treat `deinit` as 'internal by default'.
Without this, `deinit` gets included in the symbols needing documenting
as part of the public API which is unhelpful.