Pass module name from podspec to sourcekitten
Even if `--podspec` is not set, jazzy looks for a `*.podspec` file
and takes metadata from there if it is not set on the command line.
These pieces of config metadata do not show up as `_configured` so
we have to check their content for validity.

This commit restores historical behavior that was accidentally
broken in c205b2d.