Document protocol conformances from extensions
Include extensions that add protocols that are either 'public' (that is
covered by `min-acl`) or from external modules, to types that are either
'public' types from the module being documented or types from external
modules.

The ACL-related info to do this is not available from the extension
declarations, so we assume all conformances should be included and then
remove the unwanted ones after processing all the module's declarations.