Change undocumented imported extension detection
The point of this code is to not blame the user if
they write an extension of a type in another module and
that type has no documentation.

The current code guesses about the provenance of the
type based on it filesystem location.  This is fragile
(#498) and unhelpful if we don't have filesystem
locations.

The new version checks directly whether the type is from
another module -- the `key.modulename` is never set (by
design) on defs from the module being built.