Swift SymbolGraph importer (#1171)
Add `--swift-build-tool symbolgraph` to use the Swift 5.3 symbolgraph
feature to generate docs from `.swiftmodule` files.  Add to test specs and
docs.

The importer converts the symbol graph to approximately SourceKitten
output and provides that as input to the existing code.

Only intentional change to existing code is a modification of the "do we
blame the user for this declaration not being documented" question, to
deal with symbol graph output that is sometimes missing source locations.

The modification turned out to fix #498 and showed a missed 'undocumented'
declaration in a test suite.