Support pre-generated Swift symbolgraph files
Since Swift 5.5, the Swift compiler has supported the `-emit-symbol-graph`
and `-emit-symbol-graph-dir` flags to directly create Symbol graph files
as part of the compilation process. If a developer has access to these
files, it's not necessary for Jazzy to use `symbolgraph-extract` to attempt
to separately generate them.

This commit adds a `--symbolgraph_directory` configuration option to Jazzy.
If this option is provided, Jazzy skips the generation of symbolgraph files
and directly parses any `*.symbol.json` files in the given directory.