MM: Identify module source for extensions
This is to resolve the issue of merging extensions that provide
protocol compliance -- in this case there may be no members to annotate
and show the source module.

A bit messy and not entirely complete because of the way the code is
structured -- really we should rework 'declaration' to be a list of
stuff for the templating layer but it is already tangled up with html
because of Rouge.

Good enough though, covers everything, and in reality these are unusual
cases.